requirements.txt
torch
torchvision
transformers
datasets
evaluate
jiwer
Pillow
accelerate
streamlit
sentencepiece
